+++ modules/warden/src/utils/acceleration_readings.c
@@ -4,7 +4,7 @@
 #include "pico_adxl345/adxl345.h"
 
 accelerometer_reading get_accel_readings_mean(accelerometer_reading input[],
-                                               int readings_count) {
+                                              size_t readings_count) {
     accelerometer_reading means = {};
     accelerometer_reading tmp_readings = {};
 
@@ -23,7 +23,7 @@ accelerometer_reading get_accel_readings_mean(accelerometer_reading input[],
 
 void get_bunch_of_accel_readings(ADXL345I2C* adxl345_i2c,
                                  accelerometer_reading total_accel[],
-                                 int readings_count, int delay) {
+                                 size_t readings_count, int delay) {
     float current_accel_buff[3] = {0.0};
 
     for (int i = 0; i < readings_count; i++) {
